pyproject.toml
[project]
name = "ragdoc"
version = "0.0.1"
description = "Supervisor-driven RAG agents - fetch job"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.27",
  "PyYAML>=6.0.1",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "openai>=1.40",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",
  "docutils>=0.20",
  "langgraph>=0.2.34",
  "langgraph-cli>=0.3.8",
  "langchain-core>=0.2.35",
  "langchain-openai>=0.1.23",
  "langgraph-api>=0.3.1",
  "langchain-community>=0.2.15"
]

[project.scripts]
ragdoc-fetch = "ragdoc.fetch.cli:main"
ragdoc-test = "ragdoc.test.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
